requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
numpy>=1.24.0
moviepy>=2.0.0
librosa>=0.10.0
openai-whisper>=20231117
manimpango>=0.5.0
qrcode>=7.4.0
pygame>=2.5.0
opencv-python>=4.8.0
psutil>=5.9.0
pyinstaller>=6.0.0
# Separación de voces (mejor detección en canciones con beat fuerte)
demucs>=4.0.0
soundfile>=0.12.0
torch>=2.0.0
torchaudio>=2.0.0
